+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,9 @@
+Versoin X.X.X
+=============
+
+ * 2015-10-18 OptionalBinder for ObjectMapper and XmlMapper, so that users can easily override
+              the default types for both of them. (amit2103/jjlauer)
+
 Version 5.2.1
 =============